refactor: add messaging package
