+++ src/main/java/com/festp/handlers/ChatHandler.java
@@ -5,7 +5,7 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.AsyncPlayerChatEvent;
 
-import com.festp.Chatter;
+import com.festp.messaging.Chatter;
 
 public class ChatHandler implements Listener
 {
