+++ src/main/java/com/festp/handlers/SmallCommandsHandler.java
@@ -8,7 +8,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerCommandPreprocessEvent;
 import org.bukkit.event.server.ServerCommandEvent;
 
-import com.festp.Chatter;
+import com.festp.messaging.Chatter;
 
 public class SmallCommandsHandler implements Listener
 {
